requirements.txt
gymnasium==0.29.1
python-dotenv==1.1.1
requests==2.32.5
